refactor: Payment 도메인을 독립 패키지로 분리
